requirements.txt
asgiref==3.6.0
Babel==2.12.1
Django==4.2.1
django-money==3.1.0
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
py-moneyed==2.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
